gui scaling fixes (no window packing, unify methods, MML menu)
