+++ megamek/src/megamek/client/ui/swing/MegaMekGUI.java
@@ -44,7 +44,6 @@ import javax.swing.JFileChooser;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
-import javax.swing.SwingUtilities;
 import javax.swing.ToolTipManager;
 import javax.swing.UIManager;
 import javax.swing.filechooser.FileFilter;
@@ -1202,22 +1201,9 @@ public class MegaMekGUI implements IPreferenceChangeListener {
     private static void setLookAndFeel() {
         try {
             UIManager.setLookAndFeel(GUIPreferences.getInstance().getUITheme());
-            updateAfterUiChange();
+            UIUtil.updateAfterUiChange();
         } catch (Exception ex) {
             logger.error("setLookAndFeel() Exception", ex);
-        }
-    }
-
-    /**
-     * Updates all existing windows and frames. Use after a gui scale change or look-and-feel change.
-     */
-    public static void updateAfterUiChange() {
-        for (Window window : Window.getWindows()) {
-            SwingUtilities.updateComponentTreeUI(window);
-            window.invalidate();
-            window.validate();
-            window.pack();
-            window.repaint();
         }
     }
 }
